requirements.txt
python-telegram-bot
ollama
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
tinydb
pydantic
tenacity
tzdata
zoneinfo
langchain